+++ src/bas/bas_sph2ang2_ecs.F90
@@ -12,7 +12,7 @@ subroutine bas_sph2ang2_ecs(fun1, fun2)
   complex(c_double_complex), intent(in) :: fun1(1:(nrad-1), 0:lmax2, 1:nfun, 1:nfun)
   complex(c_double_complex), intent(out) :: fun2(1:(nrad-1), 1:nlat, 1:nfun, 1:nfun)
 
-  integer(c_int) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
+  integer(c_long) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
 
   !$omp parallel default(shared) private(mji, llr, ulr)
   call util_omp_disp(1, nrad - 1, llr, ulr)
@@ -50,10 +50,10 @@ subroutine bas_sph2ang2_dyn_ecs(fun1, fun2)
   complex(c_double_complex), intent(in) :: fun1(1:(nrad-1), 0:lmax2, 1:nfun, 1:nfun)
   complex(c_double_complex), intent(out) :: fun2(1:(nrad-1), 1:nlat, 1:nfun, 1:nfun)
 
-  integer(c_int) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
+  integer(c_long) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
 
 
-  integer(c_int) :: ijl, nijl, mapijl(1:3, 1:nfun*nfun*nlat)
+  integer(c_long) :: ijl, nijl, mapijl(1:3, 1:nfun*nfun*nlat)
 !  real(c_double) :: xvec(1:(nrad-1), 1:2), yvec(1:(nrad-1), 1:2)
   nijl = 0
   do ifun = nfcore + 1, nfun
@@ -93,7 +93,7 @@ subroutine bas_sph2ang2_dyn_ecs(fun1, fun2)
   !$omp end do
   !$omp end parallel
 
-!DEBUG  integer(c_int) :: ipair, npair, mappair(1:2, 1:nfun*nfun)
+!DEBUG  integer(c_long) :: ipair, npair, mappair(1:2, 1:nfun*nfun)
 !DEBUG  npair = 0
 !DEBUG  do ifun = nfcore + 1, nfun
 !DEBUG     do jfun = ifun, nfun
@@ -181,7 +181,7 @@ subroutine bas_sph2ang2_fcdyn_ecs(fun1, fun2)
   complex(c_double_complex), intent(in) :: fun1(1:(nrad-1), 0:lmax2, 1:nfun, 1:nfun)
   complex(c_double_complex), intent(out) :: fun2(1:(nrad-1), 1:nlat, 1:nfun, 1:nfun)
 
-  integer(c_int) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
+  integer(c_long) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
 
   !$omp parallel default(shared) private(mji, llr, ulr)
   call util_omp_disp(1, nradfc, llr, ulr)
@@ -227,7 +227,7 @@ subroutine bas_sph2ang2_fc1dyn_ecs(fun1, fun2)
   complex(c_double_complex), intent(in) :: fun1(1:(nrad-1), 0:lmax2, 1:nfun, 1:nfun)
   complex(c_double_complex), intent(out) :: fun2(1:(nrad-1), 1:nlat, 1:nfun, 1:nfun)
 
-  integer(c_int) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
+  integer(c_long) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
 
   !$omp parallel default(shared) private(mji, llr, ulr)
   call util_omp_disp(1, nradfc, llr, ulr)
@@ -282,7 +282,7 @@ subroutine bas_sph2ang2_fcfc_ecs(fun1, fun2)
   complex(c_double_complex), intent(in) :: fun1(1:(nrad-1), 0:lmax2, 1:nfun, 1:nfun)
   complex(c_double_complex), intent(out) :: fun2(1:(nrad-1), 1:nlat, 1:nfun, 1:nfun)
 
-  integer(c_int) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
+  integer(c_long) :: ifun, jfun, irad, ilat, l, mji, llr, ulr
 
   !$omp parallel default(shared) private(mji, llr, ulr)
   call util_omp_disp(1, nrad - 1, llr, ulr)
